@@ -1,8 +1,6 @@
 module github.com/slshen/paperscore
 
-go 1.22.0
-
-toolchain go1.22.3
+go 1.24.0
 
 require (
 	github.com/alecthomas/participle/v2 v2.1.1
