@@ -1,11 +1,15 @@
 package stats
 
 import (
+	"slices"
+
 	"github.com/slshen/paperscore/pkg/game"
 )
 
 type FieldingStats struct {
 	FieldingByPosition []*Fielding
+	PositionsByPlayer  map[game.PlayerID][]int
+	ErrorsByPlayer     map[game.PlayerID]int
 	Errors             int
 }
 
@@ -23,11 +27,24 @@ func newFieldingStats() *FieldingStats {
 	}
 	return &FieldingStats{
 		FieldingByPosition: fs,
+		PositionsByPlayer:  map[game.PlayerID][]int{},
+		ErrorsByPlayer:     map[game.PlayerID]int{},
 	}
 }
 
-func (stats *FieldingStats) recordError(e game.FieldingError) {
+func (stats *FieldingStats) recordFielder(pos int, player game.PlayerID) {
+	positions := stats.PositionsByPlayer[player]
+	if !slices.Contains(positions, pos) {
+		stats.PositionsByPlayer[player] = append(positions, pos)
+	}
+}
+
+func (stats *FieldingStats) recordError(state *game.State, e game.FieldingError) {
 	f := stats.FieldingByPosition[e.Fielder-1]
+	player := state.Defense[e.Fielder-1]
+	if player != "" {
+		stats.ErrorsByPlayer[player]++
+	}
 	f.Errors++
 	stats.Errors++
 }
